This commit added a bunch of comments and a deep but small change in code's workflow
Comments were added to the Dealer class and a single comment on the main code
was changed and added in English.s

The Ace rank value is not a list with two values anymore. The value it's 1
and will be added by 10 in case of player's choice.

There's a lot of code which needs to be changed to match this change and
do not broke the code.