Change default values at values_verification function and change main code
Most default values were deleted from values_verification function.
The function will only receive the name of who's cards are being
evaluated and his cards

Change if conditional which verify if player entered 0 in bet input was
updated to asks if player wants to keep playing, assuming he entered
0 by accident, or if he wants to stop playing. Comment was also added
in the conditional

It is assumed because if player does not have money, he won't be able
to reach this point in the code

Delete for loop which populated player's and dealer's lists and add
list comprehension insted. Added one list comprehension which generate
two lists