Refactor values_verification function completely
Change print, in case of stand, to print the sum of list_card_values
and now return this sum.

Add comments to almost all below lines of code.

Keep appending one card to list_card_values and print statement of
it's value and name.

Remove all old verifications and print statements and create a variable
called cards_values_sum to hold the list_card_values' sum up to now and
verify if this sum is equals or higher than 21. If it is, return the value.