Refactor whole values_verification function and some parts of main code
values_varification changes:
Remove while loop, hit_stand input and if conditional of 21 limite.
Add int variable to receive sum of cards values.
Add a boolean variable to confirm ace_presence in list of cards.
Add for loop to loop over all cards and get it's values.
Add value plus 10 and set ace_presence variable to true, if it is an ace.
Add normal card values if it's not an ace.
Return only cards_values_sum.

Main code changes:
Add if conditional to verify if first two cards are a blackjack.
Create and set player_value to 21, if it is and skip next while loop.
Add hit_stand input and try except block to this while loop inside main code.
Create and set player_value to 0, if it wasn't created before.
Add same for loop in values_verification funtion to sum cards numbers and
verify presence of an ace.
If player choose hit, call values_verification and add it's return value to
player_value variable.
Verify if player_value is greater or equals to 21 and break loop if it is.

Also change space between some functions and classes.
Add comments to most added new lines.