Create new sum_values function and change values_verification function
Create a new function called sum_values just to sum cards values
any time needed in the code. It its just the for loop added in the last
commit to the values_verification function and main code. This decision
was made based on the on optimization, to not repete this lines of code.

Change the if conditional of blackjack with the two first cards to insde
function values_verification.The function now call the sum_values function
at its end and return its result.

The main code was changed to use the call of sum_values instead of using
the for loop again.

Add comments to each new and changed section in the code.