Create dealer's while loop with ifs and modify some parts of main code
Create dealer's while loop to perform dealer's tasks and add if conditinals
to the loop to verify if dealer will stand or hit. The condition was based
on dealer's current cards values. If it is higher than 16 or player_value,
stand, if it isn't hit.

Readd if conditional which verify blackjack since two firsts cards
to main code because it needs to be done right after cards are showed.
Add it also to dealer.