Fix bug in the sum_values function
Creates cards_values_sum variable before loop, which, not having it,
creates a bug. This bug was not being raised because the assignment for
normal cards was using a wrong signal: '=+', insted of '+='. It was just
assigning the value withou suming it. After fixing it, the above mentioned
bug was raised.