Sum cards earlier and fix error in blackjack if
Change when cards from both player and dealer will be sumed. It happens
right after creating the lists because the value will be needed since
the cards are given.

Change the condition in the blackjack if to sum the cards values sum
plus 10 instead of trying to sum the list, which raised an error.

The dealer's condition to stop hitting at 17 was placed out while loop
because its was printing 'Total' more than one time in case dealer gets
17 or more after hit.